float64

+++ api/engine/utils.go
@@ -5,8 +5,8 @@ import (
 	"sort"
 )
 
-func mergeIngredientsWithScores(ingredients ...[]model.Ingredient) map[string]int {
-	imap := make(map[string]int)
+func mergeIngredientsWithScores(ingredients ...[]model.Ingredient) map[string]float64 {
+	imap := make(map[string]float64)
 
 	for _, ingredient := range ingredients {
 		for _, i := range ingredient {
@@ -33,8 +33,8 @@ func getIngredientsIds(is []model.Ingredient) []uint {
 	return ids
 }
 
-func IngredientsSliceToMap(is []model.Ingredient) map[string]int {
-	m := make(map[string]int)
+func IngredientsSliceToMap(is []model.Ingredient) map[string]float64 {
+	m := make(map[string]float64)
 
 	for _, i := range is {
 		m[i.Name] = i.Score
